* Make Psychtoolbox core routines and Screen() core routines   reasonably 64-bit clean.
  This should hopefully make Screen() on supported 64 bit
  runtimes safe to use with a larger than 32-bit address space.

  Memory buffers/textures/offscreen/onscreen windows etc. can
  individually and in sum exceed the 2GB limit, so usercode
  can take advantage of more than 2GB of memory.

  PTB's internal API's for passing vectors/matrices/lists from
  or the scripting environment (Matlab/Octave) into PTB still
  uses 32 bit signed integers to pass back element counts of
  vectors/matrices/arrays back to the core routines. Therefore
  the size of individual matrix dimensions etc. is still limited
  to about 2^31-1 elements, e.g., image width or height etc.
  can't exceed 2 Gigapixels in any dimension, although the total
  pixel count/component count/memory consumption of an image can
  exceed the 2 Gigaelement / 2GB limit.

  As there are various limits on max sizes imposed by current
  OpenGL API (2^31-1 or 2^32 elements), hardware (high-end GPU's
  currently 16384 x 16384 pixels) and memory, this is more of a
  theoretical limit than anything practically relevant.

  Removing the limits would be simple but tedious work, so we
  don't it until real need arises.

  The only driver which may warrant a specific api extension is
  PsychPortAudio, where PTB currently imposes a 2 Gigasamples per
  audio buffer limit, which isn't imposed by any hardware/memory
  limitations on the most high-end desktop computers.

* Remove some compiler warnings on Octave build (scalb() mismatch
  via -fno-builtin-scalb in the XCode project) and in other areas.

=> Screen for Octave and Matlab on OS/X Intel rebuilt.
=> This is lightly tested, but seems ok.
=> Let's hope this still builds on other platforms, esp. Windows.



git-svn-id: https://psychtoolbox-3.googlecode.com/svn/trunk@1991 f5efbe48-b203-0410-9429-fc5e098e5a54

+++ PsychSourceGL/Source/Common/Base/PsychConstants.h
@@ -179,7 +179,7 @@ typedef unsigned char		psych_bool;
         // Matlab 5 doesn't know about mxLOGICAL_CLASS :(
         #define mxLOGICAL_CLASS mxUINT8_CLASS
         #define mxGetLogicals(p) ((PsychNativeBooleanType*)mxGetData((p)))
-        mxArray* mxCreateNativeBooleanMatrix3D(int m, int n, int p);
+        mxArray* mxCreateNativeBooleanMatrix3D(size_t m, size_t n, size_t p);
 
         #if PSYCH_LANGUAGE == PSYCH_MATLAB
         #define mxCreateLogicalMatrix(m,n) mxCreateNativeBooleanMatrix3D((m), (n), 1)
